apps/requirements.txt
Django>=4.2, <4.3
djangorestframework>=3.14, <3.15
djangorestframework-simplejwt==5.2.2
dj-rest-auth==4.0.1
django-allauth==0.54.0
uWSGI==2.0.22
mysqlclient==2.2.0
python-dotenv==1.0.0
django-configurations==2.4.1
pytz==2023.3
drf-spectacular==0.26.4
django-cors-headers==4.2.0
django-currentuser==0.6.1
django-filter==23.2
djangorestframework-camel-case==1.4.2


# test
pytest-django==4.5.2
pytest-env==0.8.2
mock==5.1.0
factory-boy==3.3.0
parameterized==0.9.0
ipython==8.14.0
freezegun==1.2.2
tox==4.7.0

# lint
flake8==6.1.0
flake8-quotes==3.3.2
black==23.7.0
isort==5.12.0
